+++ packages/convo-lang/src/lib/convo-default-vars.ts
@@ -1,4 +1,4 @@
-import { SceneCtrl, aryRandomize, base64Encode, base64EncodeMarkdownImage, base64EncodeUrl, createJsonRefReplacer, deepClone, deepCompare, escapeHtml, escapeHtmlKeepDoubleQuote, getContentType, getErrorMessage, httpClient, joinPaths, markdownLineToString, objectToMarkdownBuffer, shortUuid, starStringTest, toCsvLines, uuid, valueIsZodObject } from "@iyio/common";
+import { SceneCtrl, aryRandomize, base64Encode, base64EncodeMarkdownImage, base64EncodeUrl, createJsonRefReplacer, deepClone, deepCompare, escapeHtml, escapeHtmlKeepDoubleQuote, getContentType, getErrorMessage, getFileNameNoExt, httpClient, joinPaths, markdownLineToString, objectToMarkdownBuffer, shortUuid, starStringTest, toCsvLines, uuid, valueIsZodObject } from "@iyio/common";
 import { vfs } from "@iyio/vfs";
 import { format } from "date-fns";
 import { ConvoError } from "./ConvoError.js";
@@ -14,6 +14,32 @@ import { convoScopeFunctionReadDoc } from "./scope-functions/convoScopeFunctionR
 const ifFalse=Symbol();
 const ifTrue=Symbol();
 const breakIteration=Symbol();
+
+const mdImg=createConvoScopeFunction(async (scope,ctx)=>{
+    const [
+        url,
+        description=getFileNameNoExt(url)?.trim().replace(/[^\w-. ]/g,'_')||'image',
+        contentType=getContentType(url)
+    ]=scope.paramValues??[];
+    try{
+        if(typeof url !== 'string'){
+            throw new ConvoError('invalid-args',{statement:scope.s},'fsReadBase64Url expects first argument to be a string');
+        }
+        if(typeof description !== 'string'){
+            throw new ConvoError('invalid-args',{statement:scope.s},'fsReadBase64Url expects second argument to be a string');
+        }
+        if(typeof contentType !== 'string'){
+            throw new ConvoError('invalid-args',{statement:scope.s},'fsReadBase64Url expects third argument to be a string');
+        }
+        const path=ctx.getFullPath(url,scope);
+
+        const data=await vfs().readBufferAsync(path);
+        return base64EncodeMarkdownImage(description,contentType,data);
+    }catch(ex){
+        console.error('ERROR',ex);
+        return ''
+    }
+})
 
 
 const mapFn=makeAnyConvoType('map',createConvoScopeFunction({
@@ -1290,7 +1316,7 @@ export const defaultConvoVarsBase={
     }),
 
     [convoFunctions.fsFullPath]:createConvoScopeFunction(async (scope,ctx)=>{
-        return ctx.getFullPath(scope.paramValues?.[0]);
+        return ctx.getFullPath(scope.paramValues?.[0],scope);
     }),
 
     [convoFunctions.joinPaths]:createConvoScopeFunction(async (scope)=>{
@@ -1462,7 +1488,7 @@ export const extendedConvoVars={
         if(typeof scope.paramValues?.[0] !== 'string'){
             throw new ConvoError('invalid-args',{statement:scope.s},'fsWriteJson expects first argument to be a string');
         }
-        const path=ctx.getFullPath(scope.paramValues[0]);
+        const path=ctx.getFullPath(scope.paramValues[0],scope);
         const value=scope.paramValues[1];
 
         await vfs().writeObjectAsync(path,value);
@@ -1473,7 +1499,7 @@ export const extendedConvoVars={
         if(typeof scope.paramValues?.[0] !== 'string'){
             throw new ConvoError('invalid-args',{statement:scope.s},'fsReadJson expects first argument to be a string');
         }
-        const path=ctx.getFullPath(scope.paramValues?.[0]);
+        const path=ctx.getFullPath(scope.paramValues?.[0],scope);
 
         try{
             return await vfs().readObjectAsync(path);
@@ -1489,7 +1515,7 @@ export const extendedConvoVars={
         }
 
         const all=await Promise.all(options.names.map(async name=>{
-            const path=ctx.getFullPath(options.pattern?options.pattern.replace('*',name):name);
+            const path=ctx.getFullPath(options.pattern?options.pattern.replace('*',name):name,scope);
             let value=await vfs().readStringAsync(path);
             if(options.tagItemsWithName){
                 value=`<${name}>\n${value}\n</${name}>`;
@@ -1512,7 +1538,7 @@ export const extendedConvoVars={
         if(typeof url !== 'string'){
             throw new ConvoError('invalid-args',{statement:scope.s},'fsReadBase64 expects first argument to be a string');
         }
-        const path=ctx.getFullPath(url);
+        const path=ctx.getFullPath(url,scope);
 
         const data=await vfs().readStringAsync(path);
         return base64Encode(data??'');
@@ -1530,39 +1556,20 @@ export const extendedConvoVars={
         if(typeof contentType !== 'string'){
             throw new ConvoError('invalid-args',{statement:scope.s},'fsReadBase64Url expects second argument to be a string');
         }
-        const path=ctx.getFullPath(url);
+        const path=ctx.getFullPath(url,scope);
 
         const data=await vfs().readBufferAsync(path);
         return base64EncodeUrl(contentType,data);
     }),
 
-    [convoFunctions.fsReadBase64Image]:createConvoScopeFunction(async (scope,ctx)=>{
-        const [url,description='image',contentType=getContentType(url)]=scope.paramValues??[];
-        try{
-            if(typeof url !== 'string'){
-                throw new ConvoError('invalid-args',{statement:scope.s},'fsReadBase64Url expects first argument to be a string');
-            }
-            if(typeof description !== 'string'){
-                throw new ConvoError('invalid-args',{statement:scope.s},'fsReadBase64Url expects second argument to be a string');
-            }
-            if(typeof contentType !== 'string'){
-                throw new ConvoError('invalid-args',{statement:scope.s},'fsReadBase64Url expects third argument to be a string');
-            }
-            const path=ctx.getFullPath(url);
-
-            const data=await vfs().readBufferAsync(path);
-            return base64EncodeMarkdownImage(description,contentType,data);
-        }catch(ex){
-            console.error('ERROR',ex);
-            return ''
-        }
-    }),
+    [convoFunctions.fsReadBase64Image]:mdImg,
+    [convoFunctions.mdImg]:mdImg,
 
     [convoFunctions.fsWrite]:createConvoScopeFunction(async (scope,ctx)=>{
         if(typeof scope.paramValues?.[0] !== 'string'){
             throw new ConvoError('invalid-args',{statement:scope.s},'fsWrite expects first argument to be a string');
         }
-        const path=ctx.getFullPath(scope.paramValues[0]);
+        const path=ctx.getFullPath(scope.paramValues[0],scope);
         const value=scope.paramValues[1];
 
         await vfs().writeStringAsync(path,(typeof value === 'string')?value:(value+''));
@@ -1573,7 +1580,7 @@ export const extendedConvoVars={
         if(typeof scope.paramValues?.[0] !== 'string'){
             throw new ConvoError('invalid-args',{statement:scope.s},'fsRead expects first argument to be a string');
         }
-        const path=ctx.getFullPath(scope.paramValues?.[0]);
+        const path=ctx.getFullPath(scope.paramValues?.[0],scope);
 
         try{
             return await vfs().readStringAsync(path);
@@ -1586,7 +1593,7 @@ export const extendedConvoVars={
         if(typeof scope.paramValues?.[0] !== 'string'){
             throw new ConvoError('invalid-args',{statement:scope.s},'fsRemove expects first argument to be a string');
         }
-        const path=ctx.getFullPath(scope.paramValues?.[0]);
+        const path=ctx.getFullPath(scope.paramValues?.[0],scope);
 
         return await vfs().removeAsync(path);
     }),
@@ -1595,7 +1602,7 @@ export const extendedConvoVars={
         if(typeof scope.paramValues?.[0] !== 'string'){
             throw new ConvoError('invalid-args',{statement:scope.s},'fsMkDir expects first argument to be a string');
         }
-        const path=ctx.getFullPath(scope.paramValues?.[0]);
+        const path=ctx.getFullPath(scope.paramValues?.[0],scope);
 
         return await vfs().mkDirAsync(path);
     }),
@@ -1604,7 +1611,7 @@ export const extendedConvoVars={
         if(typeof scope.paramValues?.[0] !== 'string'){
             throw new ConvoError('invalid-args',{statement:scope.s},'fsExists expects first argument to be a string');
         }
-        const path=ctx.getFullPath(scope.paramValues?.[0]);
+        const path=ctx.getFullPath(scope.paramValues?.[0],scope);
 
         return (await vfs().getItemAsync(path))?true:false
     }),
@@ -1613,7 +1620,7 @@ export const extendedConvoVars={
         if(typeof scope.paramValues?.[0] !== 'string'){
             throw new ConvoError('invalid-args',{statement:scope.s},'fsExists expects first argument to be a string');
         }
-        const path=ctx.getFullPath(scope.paramValues?.[0]);
+        const path=ctx.getFullPath(scope.paramValues?.[0],scope);
 
         try{
             return (await vfs().readDirAsync(path)).items.map(v=>v.path);
@@ -1705,3 +1712,4 @@ const isPrimitiveArray=(value:any)=>{
     }
     return true;
 }
+